Move components functions into services
